// src/app/model/notizia/notiziaForm.ts
/**
 * REST API
 * REST API Generation
 *
 * OpenAPI spec version: 1.0.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface NotiziaForm {
    articolo: string;
    idNotizia?: number;
    immagine?: string;
    titolo: string;
}